Use the state function as the prototype
This is funky. It's bothersome that you can't seem to really create suclasses
of Function (at least in v8). Using the original function as a prototype
at least gets us some properties (like length) that could be useful.

Anyway, it's symmetric with what I did for List with Arrays and that had
a real motivation...